requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytest
psycopg2-binary
python-dotenv
pytest
httpx
pytest-asyncio